# backend/src/utils/swagger-json.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Whatsapp-Clone - 2022",
    "description": "This is a Professional whatsapp Documentation which is designed in node js using sockets",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "kumar.divyansh312@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/users/register": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Post User Information",
        "description": "Register User",
        "operationId": "registerUser",
        "requestBody": {
          "description": "Create a new user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/userRegister"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "userRegister": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "abcd"
          },
          "password": {
            "type": "string",
            "example": "abcd"
          },
          "confirmPassword": {
            "type": "string",
            "example": "abcd"
          },
          "name": {
            "type": "string",
            "example": "abcd"
          }
        },
        "xml": {
          "name": "order"
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}